+++ src/Dap.Server/DebugAdapterServer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reactive.Disposables;
@@ -10,11 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using DryIoc;
 using MediatR;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using OmniSharp.Extensions.DebugAdapter.Protocol;
 using OmniSharp.Extensions.DebugAdapter.Protocol.Events;
 using OmniSharp.Extensions.DebugAdapter.Protocol.Models;
@@ -27,120 +22,6 @@ using OutputHandler = OmniSharp.Extensions.JsonRpc.OutputHandler;
 
 namespace OmniSharp.Extensions.DebugAdapter.Server
 {
-    public static class DebugAdapterServerServiceCollectionExtensions
-    {
-        internal static IContainer AddDebugAdapterServerInternals(this IContainer container, DebugAdapterServerOptions options, IServiceProvider outerServiceProvider)
-        {
-            container = container.AddDebugAdapterProtocolInternals(options);
-
-            if (options.OnUnhandledException != null)
-            {
-                container.RegisterInstance(options.OnUnhandledException);
-            }
-            else
-            {
-                container.RegisterDelegate(_ => new OnUnhandledExceptionHandler(e => { }), reuse: Reuse.Singleton);
-            }
-
-            container.RegisterInstance<IOptionsFactory<DebugAdapterServerOptions>>(new ValueOptionsFactory<DebugAdapterServerOptions>(options));
-
-            container.RegisterInstance(options.Capabilities);
-            container.RegisterInstance(options.RequestProcessIdentifier);
-
-            container.RegisterMany<DebugAdapterServerProgressManager>(nonPublicServiceTypes: true, reuse: Reuse.Singleton);
-            container.RegisterMany<DebugAdapterServer>(serviceTypeCondition: type => type == typeof(IDebugAdapterServer) || type == typeof(DebugAdapterServer), reuse: Reuse.Singleton);
-
-            // container.
-            var providedConfiguration = options.Services.FirstOrDefault(z => z.ServiceType == typeof(IConfiguration) && z.ImplementationInstance is IConfiguration);
-            container.RegisterDelegate<IConfiguration>(_ => {
-                    var builder = new ConfigurationBuilder();
-                    if (outerServiceProvider != null)
-                    {
-                        var outerConfiguration = outerServiceProvider.GetService<IConfiguration>();
-                        if (outerConfiguration != null)
-                        {
-                            builder.AddConfiguration(outerConfiguration, false);
-                        }
-                    }
-
-                    if (providedConfiguration != null)
-                    {
-                        builder.AddConfiguration(providedConfiguration.ImplementationInstance as IConfiguration);
-                    }
-
-                    return builder.Build();
-                },
-                reuse: Reuse.Singleton);
-
-            return container;
-        }
-
-        public static IServiceCollection AddDebugAdapterServer(this IServiceCollection services, Action<DebugAdapterServerOptions> configureOptions = null)
-        {
-            return AddDebugAdapterServer(services, Options.DefaultName, configureOptions);
-        }
-
-        public static IServiceCollection AddDebugAdapterServer(this IServiceCollection services, string name, Action<DebugAdapterServerOptions> configureOptions = null)
-        {
-            // If we get called multiple times we're going to remove the default server
-            // and force consumers to use the resolver.
-            if (services.Any(d => d.ServiceType == typeof(DebugAdapterServer) || d.ServiceType == typeof(IDebugAdapterServer)))
-            {
-                services.RemoveAll<DebugAdapterServer>();
-                services.RemoveAll<IDebugAdapterServer>();
-                services.AddSingleton<IDebugAdapterServer>(_ =>
-                    throw new NotSupportedException("DebugAdapterServer has been registered multiple times, you must use DebugAdapterServer instead"));
-                services.AddSingleton<DebugAdapterServer>(_ =>
-                    throw new NotSupportedException("DebugAdapterServer has been registered multiple times, you must use DebugAdapterServer instead"));
-            }
-
-            services
-                .AddOptions()
-                .AddLogging();
-            services.TryAddSingleton<DebugAdapterServerResolver>();
-            services.TryAddSingleton(_ => _.GetRequiredService<DebugAdapterServerResolver>().Get(name));
-            services.TryAddSingleton<IDebugAdapterServer>(_ => _.GetRequiredService<DebugAdapterServerResolver>().Get(name));
-
-            if (configureOptions != null)
-            {
-                services.Configure(name, configureOptions);
-            }
-
-            return services;
-        }
-    }
-
-    public class DebugAdapterServerResolver : IDisposable
-    {
-        private readonly IOptionsMonitor<DebugAdapterServerOptions> _monitor;
-        private readonly IServiceProvider _outerServiceProvider;
-        private readonly ConcurrentDictionary<string, DebugAdapterServer> _servers = new ConcurrentDictionary<string, DebugAdapterServer>();
-
-        public DebugAdapterServerResolver(IOptionsMonitor<DebugAdapterServerOptions> monitor, IServiceProvider outerServiceProvider)
-        {
-            _monitor = monitor;
-            _outerServiceProvider = outerServiceProvider;
-        }
-
-        public DebugAdapterServer Get(string name)
-        {
-            if (_servers.TryGetValue(name, out var server)) return server;
-
-            var options = name == Options.DefaultName ? _monitor.CurrentValue : _monitor.Get(name);
-
-            var container = DebugAdapterServer.CreateContainer(options, _outerServiceProvider);
-            server = container.Resolve<DebugAdapterServer>();
-            _servers.TryAdd(name, server);
-
-            return server;
-        }
-
-        public void Dispose()
-        {
-            foreach (var item in _servers.Values) item.Dispose();
-        }
-    }
-
     public class DebugAdapterServer : JsonRpcServerBase, IDebugAdapterServer, IInitializeHandler
     {
         private readonly DebugAdapterHandlerCollection _collection;
